fix(zkevm-circuits): enable tx_id_diff check iff q_enable is 1
In PR #11, the calldata fixed column was set to all 1's
for the maximum call size. As a result, the `q_enable` became 0 and
`is_calldata` became 1 in certain cases, which previously did not occur.
This introduced a new bug in the tx_circuit lookup. This commit fixes that bug.